% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Edu3D}
\alias{Edu3D}
\title{Edu3D}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 119 columns.}
\usage{
Edu3D
}
\description{
20 X 119 Discovery matrix of testing a 3D educational software
}
\keyword{datasets}
