% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Infpump}
\alias{Infpump}
\title{Infpump}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 107 rows and 34 columns.}
\usage{
Infpump
}
\description{
34 X 107 Discovery matrix of testing a medical infusion pump
}
\keyword{datasets}
