% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Macerr}
\alias{Macerr}
\title{Macerr}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 145 rows and 15 columns.}
\usage{
Macerr
}
\description{
15 X 145 Discovery matrix
}
\keyword{datasets}
