% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Mantel}
\alias{Mantel}
\title{Mantel}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 30 rows and 76 columns.}
\usage{
Mantel
}
\description{
76 X 30 Discovery matrix
}
\keyword{datasets}
