% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Savings}
\alias{Savings}
\title{Savings}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 44 rows and 34 columns.}
\usage{
Savings
}
\description{
34 X 44 Discovery matrix
}
\keyword{datasets}
